Breakout individual task cards to own components
